requirements.txt
# Requirements file for the production ec2 instance

# Checking Python dependencies
piprot==0.9.6

# Using APIs
requests==2.10.0

# Checking ssl expiration
pyOpenSSL==16.0.0

# HTTP Observatory scanner
httpobs-cli==1.0.2
